refactor: 이미지 업로드 로직 helper로 분리
