+++ feature/src/main/java/org/maru/muaring/feature/member/ui/ProfileSetupFragment.java
@@ -10,8 +10,6 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageButton;
-import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -22,6 +20,7 @@ import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 import org.maru.muaring.core.ui.ImagePickerView;
+import org.maru.muaring.data.helper.ProfileSetupState;
 import org.maru.muaring.feature.R;
 import org.maru.muaring.feature.common.navigation.CommonNavigator;
 import dagger.hilt.android.AndroidEntryPoint;
